+++ gfx/fonts/gl_font.h
@@ -21,7 +21,7 @@
 
 typedef struct gl_font_renderer
 {
-   bool (*init)(void *data, const char *font_path, unsigned font_size);
+   bool (*init)(void *data, const char *font_path, float font_size);
    void (*deinit)(void *data);
    void (*render_msg)(void *data, const char *msg);
    void (*render_msg_place)(void *data, float x, float y,
@@ -33,7 +33,7 @@ extern const gl_font_renderer_t gl_raster_font;
 extern const gl_font_renderer_t libdbg_font;
 
 const gl_font_renderer_t *gl_font_init_first(void *data,
-      const char *font_path, unsigned font_size);
+      const char *font_path, float font_size);
 
 #endif
